requirements.txt
requests
pandas
beautifulsoup4
lxml
flask
flask-login
gunicorn
psycopg2-binary
